pyproject.toml
[project]
name = "portkey-guardrail"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "modal",
    "portkey-ai",
    "instructor",
    "ipython>=9.9.0",
    "datasets>=4.5.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "ssort>=0.16.0",
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.13",
]
